+++ pyrate_limiter/__init__.py
@@ -8,3 +8,4 @@ from .constants import *
 from .exceptions import *
 from .limiter import *
 from .request_rate import *
+from .sqlite_bucket import *